py: 优化流的超时处理

+++ python/socketd/socketd/transport/client/ClientSession.py
@@ -25,11 +25,11 @@ class ClientSession:
         ...
 
     @abc.abstractmethod
-    def send_and_request(self, event: str, content: Entity, timeout: int = 0) -> RequestStream:
+    def send_and_request(self, event: str, content: Entity, timeout: float = 0) -> RequestStream:
         ...
 
     @abc.abstractmethod
-    def send_and_subscribe(self, event: str, content: Entity, timeout: int = 0) -> SubscribeStream:
+    def send_and_subscribe(self, event: str, content: Entity, timeout: float = 0) -> SubscribeStream:
         ...
 
     @abc.abstractmethod
